Follow neutron's master
Our master branch should now follow Neutron's master branch.

For this, a significant refactoring of bagpipe linuxbridge
agent was necessary, which happens to make he code a bit
simpler/cleaner: in particular, the RPCs are now registered
through an agent extension, and we are closer to being able
to simply run as a linuxbridge agent extension.

We also need to follow the movement of some definitions
from neutron to neutron-lib, this patch adds the appropriate
dependency and accordingly updates the code.

This change also makes bagpipe-bgp client ready for being
used in the networking-bgpvpn bagpipe driver OVS agent
extension, by changing the bagpipe-bgp client code so
that it can simply receives two OVS bridge objects for
tun_br and int_br, and use these instead of parameters
passed at init:
- find the ofport for the tun/int patch port
- find the local vlan for a said network/port
- setup ARP entries

Change-Id: Ief6cdaab338830efc37827b946700443a1d6b64e