Use patch-port:vlan i.o. linux interface as linuxif
Providing the linuxinterface (qvoxxxx) as 'linuxif' to bagpipe-bgp
is not as harmless as it should be, because this is the key used
to determine when the last plug on an interface is done.

This patch provides the patch-port:vlan tuple instead, which should
allow bagpipe-bgp current code to properly identify the unplug of the
last VM on a Network on a given compute node.

Change-Id: I1f38815e188b4d0cd72b15d66319194fcfce1e0a
Closes-Bug: 1644151
(cherry picked from commit e715f1531d1bd2c776a8d1cf02a80ffb80ebb25e)