bagpipe-bgp, bare MPLS: ARP trigger on specific interface
When abre MPLS is used, ARP is currently triggered on the
interface that routes to the BGP next hop.

In practice, the interface has to be the MPLS bridge.

This change modifies the driver behavior and the devstack
plugin system configuration function to reflect that.

Additionally, a ping is used just to trigger an ARP resolution,
but we don't need the ping to actually succeed, and we want
in fact to support cases where the neighbor may not be responding
to pings. So we proceed even if the ping failed.

Change-Id: Ia7a0821618f4f7a56596b90a9894a6a92618226f