Automatically generate configuration files
Auto-generation of configuration files is part of the community-wide
doc-migration effort in Pike. networking-bagpipe does not support it,
so this commit sets up the config file autogeneration.

Note that lower case is used as config group names as oslo.config suggests
to do so and oslo.config sphinx extensions expects to use lower cases
(otherwise it emits warnings).

Change-Id: I69abd82cd5e422ec2d52f8d82e45a141c05bdb6a