bagpipe-bgp: OVS, avoid use of two patch ports
Original code for integrating br-mpls and br-tun was
using two patch ports, one for each direction of traffic,
to work around the fact that OVS/Openflow has a check to
prevent sending a packet back from the port via which it
came in.

We weren't aware of a simpler solution provided by OVS to
workaround this check, consisting in zeroing the value of
in_port with "load:0->NXM_OF_IN_PORT[]".

(see http://docs.openvswitch.org/en/latest/faq/openflow/
"Q: I added a flow to send packets out the ingress port...")

This change implements this simpler workaround to avoid the
hassle and complexity of yet another patch port.

Change-Id: I3defa9ec2cb85a6f0c902086970c6c3bf9ecd14e