bagpipe-bgp: remove tracker_worker verbosity
This change really makes TrackerWorker processing less verbose.

Since tracker work delegates logging to children class, adjusting
this loglevel for the module was ineffective.

This change relies on the use of the TRACE log level instead.

To have TrackerWorker be verbose, it will be needed to set the default_log_level
for the networking_bagpipe.bagpipe_bgp.vpn.ipvpn or
networking_bagpipe.bagpipe_bgp.vpn.evpn modules to TRACE.

Change-Id: Id74d8248c01ae698c04c467ff05b4cf748c67775