bagpipe-bgp translate_api_internal fix
The code was iterating over a dict keys and in the same loop
adding/removing records in the same dict.

This is (a) not a good idea in the first place and (b) the cause
for bug #1703382. It only shows up under python3, because
dict.keys() is a view in python 3.

What remains obscure to me is that this code was *not* causing
Python to complain with a "RuntimeError: dictionary changed
size during iteration'" exception.

This change fixes the code to avoid using an iterator for keys,
and also update the corresponding unit test, for style only.

Change-Id: I377e0449242c9495f237b52c26fe91f6b6b739f8
Closes-Bug: 1703382