gate fullstack hook: have tox and zuul-cloner inherit all environment
Pass all current environment to tox, so that ZUUL_REF in particular
is seen by tox. This seems required to have zuul Depends-On magic
work as expected.

This change also tweaks gate hooks and tox_install.sh so that we can
see more tox and zuul-cloner logs in the logs retrieved by OpenStack infra.

Change-Id: I7c1690ed6e8b4fd94bdc91fade4c3d14fc0eb14a