IPVPN OVS driver: longest match lookups
To implement longest match IP lookups, the OVS flows
for IP lookups are modified to use a dynamic priority
computed based on prefix length (longest prefixes get
higher priority).

This change introduces constants for all priorities,
including for the priority of fallback rules which are
not related to longest match lookups.

Closes-Bug: 1714524
Change-Id: I21f7a12df371bf9989febac31d56e8a1ece56f5c