bagpipe-bgp: improve TrackerWorker and RTM testing
Some code refactoring in TrackerWorker and RouteTableManager,
in particular to prepare a proper fix to bug bug/1716893 and
avoid TrackerWorker and RouteTableManager tests to change threads
(which makes it a pain to debug/trace with pdb).

The name of callbacks for TrackerWorker subclasses are
changed to avoid a leading _ because it's not natural to
consider these private methods.

Change-Id: Iec2428c89419293687aee6ebe398e430b4816793