bgpvpn agent extension: fix disassociate after update
We need to ignore the 'id' attribute in the RPC dict
for update_bgpvpn and delete_bgpvpn. It is a remain of
past code, and useless, and it breaks the _is_last_bgpvpn_info
method.

This change also updates existing unit tests so that
we emulate the reception of real RPCs from the neutron server
for bgpvpn.

Closes-Bug: 1722569
Change-Id: I7c479ff60328c02761ac865bb11318b52f1427ec