BGPVPN L2/EVPN: use same VNI as the associated network
With this change, the following happens when a VXLAN network is
associated to an L2 BGPVPN: bagpipe-bgp will be given, as a VNI
to use for the EVPN, the VNI of the network.  This currently applies
to linuxbridge only (no EVPN OVS driver yet).

Additionally, the behavior of bagpipe-bgp EVPN/VXLAN driver is modified
so that whenever a VXLAN interface with the target VNI already exists
on the bridge for a given EVPN, this interface will be reused.

The result is that when a VXLAN network is associated to an L2 BGPVPN,
the VNI of the network will be used for EVPN and the VXLAN interface
of the bridge for this network will be reused, avoiding tricky questions
that would be raised if two vxlan interfaces were used on that bridge.

To make this possible, changes were brought around agent extensions to
adadpt to the fact that the VNI of a network is known by our extension
at handle_port time, which may be after the reception of the
bgpvpn_port_attach RPC.

Change-Id: I9b360a90826104b09462542483f0b08df632531c