bagpipe-bgp: add cleanup assist
The cleanup assist mecanism is introduced to handle the
case where at stop/cleanup time, a dataplane driver has
no simple mean to remove all the states it has created
in the dataplane.

We enhance VPNInstance and TrackerWorker so that, on stop(),
if the dataplane driver requires it, we simulate the withdrawal
of all received routes.

This is *not* done based on the unsubscribe mechanism, because
doing so would run into the problem of having to identify the
moment where all routes have been withdrawn, which the
RouteTableManager does not support.

This enhancement is used by the EVPN/VXLAN linux driver for the
cases where the vxlan interface is not created by the driver, but
reused from another component, making it impossible to remove
all dataplane states by removing the vxlan interface.

This change also replace the VPNInstance Lock with an RLock
which really is the kind of Lock we need.

Change-Id: I8a00dc0a98f52be022a938254062db0abe43991f