EVPN/VXLAN linux: fix unplug
The change in Ie33880f1d6cca10a391bbe6a7cb13597046f597b does not
play out nicely with the use of bagpipe-bgp with linuxbridge agent:
on unplug it seems that the tap interface has already been removed.

This change adds a check before removing the state that the VM
interface is still on the bridge (if it's not, no cleanup is needed).

Change-Id: If6a85ffa7c4bc2bd837317eeebddbcb31774d081