bagpipe-bgp: EVPN/VXLAN interop fix
We need to interpret the content of EVPN routes based on whether the
dataplane is VXLAN or not: if it is, then we need to read the raw value,
not the 4-bit shifted value that applies for an MPLS label.

Similarly in the routes we advertise, we must avoid 4-bit shifting the
VXLAN VNI value.

These changes require modification in ExaBGP encoding/decoding classes,
and result in requiring ExaBGP > 4.0.2 (unreleased yet) [1].

[1] https://github.com/Exa-Networks/exabgp/pull/744

Closes-Bug: 1732668
Change-Id: I3e73ed30dc2936820a50518a1d7a18d34f5a0e65