agent extensions: remove 'static_routes' special casing
With this change, services produce attachment information
as raw as possible. In particular, a service can produce
multiple attachments.

This changes how so-called 'static routes' are handled,
instead of adding a new 'static_routes' key in attachment
information, a service will produce one attachment for
each static route.

This also allows to fix bug 1744241: the 'linuxif' key
provided to bagpipe-bgp when unplugging for an IPVPN with
linuxbridge was wrong. This fix is not in a separate change,
being hard to address without this change.

Closes-Bug: 1744241
Change-Id: I621494372ff91fbb7b8f811134fcb1dafdb915e8