bagpipe-bgp: properly withdraw route on detach
If advertise_subnet was not provided on detach, the BGP
route withdrawal was done with a different RD than the announce,
and hence failing.

This change stores the route advertised on vif_plugged to
reuse the exact same route on vif_unplugged to avoid this
issue and similar ones.

(this change also restablish code in bagpipe_bgp_agent
that was mistakenly reverted in
67228906e8ee89186c3e8c7869df91758027f0a3)

Change-Id: I9c232656814e6e8b7d271b498c6907d9cf42dfb8
Closes-Bug: 1743574