bagpipe-bgp: improve update of export RTs
The current behavior, on an update of export RTs
was to re-originated all recevied routes with the new
export RTs, includings for re-advertised routes for which
the export RTs are not the RTs that should be used
(it should be readvertise/to_rts).  This problem was not
exposed in the use made of readvertisements by BGPVPN
and SFC drivers.

Another issue is that there was a concurrency issue
between VPNInstance reading the list of received routes,
and RouteTableManager populating this list, resulting
in rare occurences of such a traceback:

Traceback (most recent call last):
  File "/opt/stack/new/networking-bagpipe/networking_bagpipe/bagpipe_bgp/api/controllers.py", line 210, in process
    self.manager.plug_vif_to_vpn(**attach_params)
  File "/opt/stack/new/networking-bagpipe/networking_bagpipe/bagpipe_bgp/common/log_decorator.py", line 40, in wrapper
    return method(*args, **kwargs)
  File "/opt/stack/new/networking-bagpipe/networking_bagpipe/bagpipe_bgp/vpn/manager.py", line 338, in plug_vif_to_vpn
    vpn_instance.update_route_targets(import_rts, export_rts)
  File "/opt/stack/new/networking-bagpipe/networking_bagpipe/bagpipe_bgp/common/log_decorator.py", line 40, in wrapper
    return method(*args, **kwargs)
  File "/opt/stack/new/networking-bagpipe/networking_bagpipe/bagpipe_bgp/vpn/vpn_instance.py", line 458, in update_route_targets
    for route_entry in self.get_route_entries():
Set changed size during iteration

Change-Id: Ie55b29f4620c99e9ba3ba1c402ed0f7cf1cc9e98