bagpipe ML2: reuse network VNI
Since Queens, bagpipe ML2 uses the vxlan type driver and hence ought
to use as a VXLAN ID the VNI allocated by neutron.

However the current code does not pass the VNI from the bagpipe ML2
agent extension to bagpipe-bgp, which results in:
- allocating a different/new VXLAN VNI
- creating a new vxlan interface on the bridge (with linuxbridge)
  rather than reusing the interface created by the Neutron agent,
  resulting in dataplane loops

With this change the VNI is correctly passed to bagpipe-bgp (which
allows bagpipe-bgp to reuse the existing VXLAN interface).

Change-Id: I11c8bfa70a048dbf8a8cf0dffef302f7efd98771
Closes-Bug: 1756361