bagpipe-bgp: strictly avoid having same VNI in two VPN instances
In some BGPVPN interconnection scenarios, two EVPNs may be
configured for bagpipe-bgp with same VNI.

This can't be supported by the linux vxlan driver (only a
single vxlan interface can be created for a given VXLAN id,
and it can't be plugged in more than one bridge).

This would be supported by the EVPN OVS driver with some
change.

For these reason, the right thing is to reject the creation
of an EVPN instance having same VNI than another instance.

Closes-Bug: 1765764

Change-Id: Ia36c0d17031042bada1df6945573cb2dabf5560a