Upgrade pylint to a version that works with python3
Fixes a few pylint warning from this recent pylint.

len-as-condition, inconsistent-return-statements and
no-else-return rules are
added to the ignore list in .pylintrc because they are
patterns commonly used in the neutron stadium and
already ignored in the neutron repo.

astroid version needs to be specified too.
Otherwise, a newer version of astroid which conflicts with pylint 2.2.0
will be installed. astroid and pylint are tightly coupled,
so both versions should be specified.

urllib2 is only available for python2.
In python3, it has been merged into urllib.
six.moves.urllib needs to be used.

Change-Id: Ib8f25b8103f6bb429cf5fbaa7a071dab7326930f
Co-Author: Akihiro Motoki <amotoki@gmail.com>