fix lower constraints and add required projects
Lower constraints jobs are failing when using depends-on with a
neutron patch due to ryu lower constraint not matching what neutron
needs for cffi [1].

This patch removes cffi from the lower-constraints as ryu is not
required directly by this project.

In addition this patch adds the required-projects to the legacy jobs
in zuul.yaml (tempest-dsvm-bgpvpnn-bagpipe and fullstack jobs)
as they are not pulling the required projects from source
and hence getting stale code.

[1] http://logs.openstack.org/22/636422/1/check/openstack-tox-lower-constraints/
26be9d9/job-output.txt.gz#_2019-02-12_19_24_08_303848

Change-Id: Ib57a6093ab0de7e9668d29f638ff5352d5106e44