Update lower-constraints to valid with recent changes in Neutron
Because of recent changes in Neutron, like remove of_interface config
option or tranition from ryu to os-ken project, it is needed to
bump minumum required versions of projects like neutron for
networking-bagpipe.

This commit updates also versions of other libs to be the same as are
configured in neutron.

This commit updates also sphinx requirements to be the same as for
other OpenStack projects.

Change-Id: Ibf46b4bbda9b44586d236752a2bad09820dd3761