PDF documentation build
README.rst is simplified and most of the contents of README is
moved to overview.rst. This is mainly to aviod inappropriate sections
in PDF doc, but README file is too long considering we have a separate
documentation so I believe this makes sense.

The content of the top page is also adjusted to
fit both HTML and PDF docs.

Change-Id: I43dbe8cd112b93d1300ddef3df8eae52c00e46b4
Story: 2006099
Task: 35121