bagpipe-bgp: cleanly ignore RTC route of unsupported type
When a BGP peer advertises an RTC route (RFC4684) with
a route target of a type bagpipe-bgp does not support, we
need to ignore these cleanly, and not break and reset the
BGP peering.

Note that it is safe to plainly ignore such routes because,
since bagpipe-bgp does not support them, it will never need
to propagate event for such RTs to a BGP peer.

Change-Id: Ie8d441b9d7669ec22236b2273d8e69a884751309
(cherry picked from commit f4113085b069683f851e475c8bc738ce4d7478b2)