Fix gate failure
When creating a tox env, networking-bgpvpn train version is installed.
It has a dependency of horizon<17.0.0 (train) which requires Django<2.1
and it is not compatible with u-c and horizon ussuri. It leads to a tox
creation failure. We need to install ussuri version of networking-bgpvpn.
A beta1 of networking-bgpvpn has been released, so we can use it.

neutron-tempest-plugin-bgpvpn-bagpipe is non-voting temporarily
due to the bagpipe-bgp CLI change. It will be fixed by
https://review.opendev.org/#/c/708648/ and we can make it voting again.

Change-Id: I81acd61f43221dd59c9c5220b989f4a9fcf06564