[stein-only] Fix networking-bgpvpn and neutron requirement constraints
Ussuri version of networking-bgpvpn (12.0.0) was released and with that
version of networking-bgpvpn tox -e docs started to fail when tox
installs latest bgpvpn. Let's constrain networking-bgpvpn to Stein
versions. This needs to be done in requirements.txt, since
networking-bgpvpn is not part of upper-constraints.txt.

Now that requirements.txt is modified the check job finds problem with
neutron constraints as well. The problems comes due to neutron was
meanwhile added to upper-constraints.txt. This patch fixes neutron's
requirements to make the check job pass.

Docs tox target's dependencies had to be updated to make use of the
upper-constraints.txt. However sphinx installs bagpipe and its
requirements.txt it does not use the upper constraints so let's
install the requirements earlier with the 'deps'.

Depends-On: https://review.opendev.org/#/c/730379
Change-Id: I599ba9823bbf3abc770cb9893489f42c7dbf477e