[CI] Remove periodic stable jobs from unmaintained branch
Periodic stable jobs are meant for stable branches and not for
unmaintained branches. perioid-weekly is enough (if there is
anyone, who checks its result) to keep. This patch removes all
periodic-stable jobs.

This patch also removes networking-sfc from required projects in
zuul.yaml as unmaintained/zed and older branches do not exist
anymore for this repository. Furthermore, in this branch we have
to cap as well networking-sfc to avoid incompatibilities with
current neutron version.

Lower constraints are not used and tested anymore in most of the
OpenStack projects so it doesn't make sense to maintain it here,
hence the job is removed and the lower-constraints.txt as well.

Also include this patch to unblock the gate: Cap setuptools <71.0.0

py39 jobs (on ubuntu-focal) started to fail (see the relevant github
issue [1]) due to recent virtualenv release (20.26.4; which bundles
setuptools).

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

[1] https://github.com/pypa/setuptools/issues/4483

Conflicts:
  .zuul.yaml

Change-Id: I085f3343e1466f43e4318055e252347000397f0f
(cherry picked from commit 79b2d924c9baeb4d824882c0473f0ca0060832cb)
(cherry picked from commit 4570f681ec7873d854dbca32508008c292315491)